# application/schema/local-land-charge-swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Local Land Charge API",
        "description": "Land Registry Local Land Charge API",
        "version": "0.0.1"
    },
    "host": "local-land-charge",
    "schemes": [
        "https"
    ],
    "basePath": "/api.landregistry.gov.uk/",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/records/": {
            "post": {
                "summary": "Add a new Local Land Charge",
                "operationId": "addLocalLandCharge",
                "description": "Add a Local Land Charge to the Local Land Charge Register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Local-Land-Charge"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "URL to the GET endpoint for the Local Land Charge is returned on successful creation.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request due to invalid schema. Response will include 1 or more schema errors",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/record/{primary_id}": {
            "get": {
                "summary": "Get the Local Land Charge record",
                "description": "The Local Land Charge endpoint returns details of a specific local land charge record based on the unique primary id.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "primary_id",
                        "in": "path",
                        "description": "Unique reference of the Local Land Charge record.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "LocalLandCharge"
                ],
                "responses": {
                    "200": {
                        "description": "A specific Local Land Charge record is returned",
                        "schema": {
                            "$ref": "#/definitions/Local-Land-Charge"
                        }
                    },
                    "404": {
                        "description": "Local Land Charge not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update the Local Land Charge record.",
                "description": "Make an update to the Local Land Charge record of the given primary id.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "primary_id",
                        "in": "path",
                        "description": "Unique reference of the Local Land Charge record.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Local-Land-Charge"
                        }
                    }
                ],
                "tags": [
                    "LocalLandCharge"
                ],
                "responses": {
                    "200": {
                        "description": "The Local Land Charge record is returned with the updated data.",
                        "schema": {
                            "$ref": "#/definitions/Local-Land-Charge"
                        }
                    },
                    "400": {
                        "description": "Bad Request due to invalid schema. Response will include 1 or more schema errors.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Local Land Charge not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/records/geometry/{geometry-function}": {
            "post": {
                "summary": "Search for Local land charges via a spatial search",
                "operationId": "searchLocalLandCharge",
                "description": "Search for Local land charges via a spatial search. {geometry-function} is a PostGIS functions but please use 'intersects' for this function",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Search-Criteria"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a collection of local land charges as defined by #/definitions/Local-Land-Charge",
                        "schema": {
                            "type": "application/json"
                        }
                    },
                    "400": {
                        "description": "Bad Request due to invalid schema. Response will include 1 or more schema errors",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CURIE": {
            "description": "Abbreviated syntax for a URI, consisting of the name of the register followed by a colon, and then the primary identifier of the record. For example 'local-land-charge:123'",
            "type": "string",
            "pattern": "\\S+:\\S+"
        },
        "OldRegisterPart": {
            "description": "The part of the Local Land Charges register that a Local Land Charge entry resides in prior to migration",
            "type": "string",
            "enum": [
                "1",
                "2",
                "3b",
                "4",
                "6a",
                "6b",
                "7",
                "8",
                "9",
                "10",
                "11",
                "12"
            ]
        },
        "Geometry": {
            "type": "object",
            "description": "Spatial representation of the Local Land Charge. Can be a point, line polygon or a collection of geometry types.",
            "properties": {
                "geometry": {
                    "type": "string"
                }
            }
        },
        "FurtherInformationDetails": {
            "type": "object",
            "required": [
                "information-location"
            ],
            "properties": {
                "information-location": {
                    "type": "string",
                    "description": "The location where a Local Land Charge document is held, physical or digital.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/CURIE"
                        }
                    ]
                },
                "references": {
                    "type": "array",
                    "maxItems": 3,
                    "items": {
                        "type": "string",
                        "pattern": "\\S+"
                    }
                }
            }
        },
        "FurtherInformation": {
            "type": "array",
            "description": "The place (physical (mandatory) and virtual (optional)) where the relevant documents pertaining to a local land charge registration may be inspected and/or copies obtained.",
            "minItems": 1,
            "maxItems": 3,
            "items": {
                "$ref": "#/definitions/FurtherInformationDetails"
            }
        },
        "Local-Land-Charge": {
            "type": "object",
            "required": [
                "geometry",
                "charge-description",
                "charge-type",
                "further-information",
                "originating-authorities"
            ],
            "properties": {
                "local-land-charge": {
                    "type": "string",
                    "description": "NOTE: Unique identifier for this charge. Do not provide when creating a charge (POST request). Only required for PUT requests"
                },
                "geometry": {
                    "$ref": "#/definitions/Geometry"
                },
                "registration-date": {
                    "description": "Date of entry on the Land Charges Register, for existing charges it is the date entered on the current register",
                    "type": "string",
                    "format": "date"
                },
                "charge-description": {
                    "description": "A description of the land affected by the LLC in terms of its geographic location - i.e. it describes that depicted by an associated Polygon.",
                    "type": "string",
                    "pattern": "\\S+"
                },
                "statutory-provisions": {
                    "type": "array",
                    "description": "The statutory provision(s) under which a local land charge has arisen, whether expressly (e.g. under s224 Highways Act 1980 or s106 Town and Country Planning Act 1990) or otherwise (e.g. conditional planning permission granted under s70 Town and Country Planning Act 1990).",
                    "maxItems": 3,
                    "items": {
                        "$ref": "#/definitions/CURIE"
                    }
                },
                "charge-type": {
                    "description": "The general category of charge (e.g. Planning, Financial). A description of the general heading under which type of charge the specific charge falls. NB: For extant / migrated charges this is not necessarily the same as the part in which it has actually been registered, although in the majority of cases it will be.",
                    "type": "string",
                    "pattern": "\\S+"
                },
                "migrating-authority": {
                    "description": "Local Authority from which and at which the Local Land Charge was migrated. Legacy field, not required for new Local Land Charge register.",
                    "type": "string",
                    "pattern": "\\S+"
                },
                "old-register-part": {
                    "$ref": "#/definitions/OldRegisterPart"
                },
                "further-information": {
                    "$ref": "#/definitions/FurtherInformation"
                },
                "creation-date": {
                    "description": "The date a Local Land Charge first came into effect. This could be the date of the instrument, the date of the decision etc.",
                    "type": "string",
                    "pattern": "^\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])$",
                    "format": "date"
                },
                "expiration-date": {
                    "description": "The date on which a Land Interest in respect of a local land charge ceases to exist.",
                    "type": "string",
                    "pattern": "^\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])$",
                    "format": "date"
                },
                "instrument": {
                    "description": "The document giving rise to a Local Land Charge e.g. charge, deed, agreement etc.",
                    "type": "string",
                    "pattern": "\\S+"
                },
                "originating-authorities": {
                    "type": "array",
                    "description": "The Minister of the Crown, government department, local authority or other person by whom the charge is brought into existence or by whom, on its coming into existence, the charge is enforceable.",
                    "items": {
                        "$ref": "#/definitions/CURIE"
                    }
                },
                "unique-property-reference-numbers": {
                    "description": "Unique Property Reference Numbers",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "land-description": {
                    "description": "A verbal description of the land adjoining the land compulsorily purchased for public works.",
                    "type": "string",
                    "pattern": "\\S+"
                },
                "works-particulars": {
                    "description": "A verbal description of the reason why the land adjoining the retained land was compulsorily acquired by an authority with compulsory purchasing powers - the particulars and extent of the works.",
                    "type": "string",
                    "pattern": "\\S+"
                },
                "capacity-description": {
                    "description": "A description of the capacity in which the land was acquired; e.g. freehold.",
                    "type": "string",
                    "pattern": "\\S+"
                },
                "compensation-paid": {
                    "description": "Amount already paid to the landowner from which the land was acquired in respect of damages to his retained land and the total sum agreed.",
                    "type": "string",
                    "pattern": "\\S+"
                },
                "archived": {
                    "description": "Boolean to indicate whether the record has been archived.",
                    "type": "boolean"
                }
            }
        },
        "Search-Criteria": {
            "type": "object",
            "required": [
                "geometry"
            ],
            "properties": {
                "geometry": {
                    "$ref": "#/definitions/Geometry"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}